move miner_ask and storage_miner objects into model package
